moved the Practice2 file from components to templates folder
